Switch KisPainter to use floating point opacity instead of 0...255 integer range
Globally, we store opacity as a floating point value. It was quite weird that we
converted the value to quint8, when passing to KisPainter, then back to float when
passing to KoCompositeOp and then, finally, to channels_type.

Now KisPainter operates with floating-point opacity value. It removes all these
extra conversions and errors when rounding.

BUG:356462

# Conflicts:
#	libs/ui/processing/KisEncloseAndFillProcessingVisitor.cpp
#	libs/ui/processing/KisEncloseAndFillProcessingVisitor.h
#	plugins/tools/tool_enclose_and_fill/KisToolEncloseAndFill.cpp

+++ libs/ui/processing/KisEncloseAndFillProcessingVisitor.h
@@ -41,7 +41,7 @@ public:
         bool unmerged,
         bool useBgColor,
         bool useCustomBlendingOptions,
-        int customOpacity,
+        qreal customOpacity,
         const QString &customCompositeOp
     );
 
@@ -73,7 +73,7 @@ private:
     bool m_unmerged {false};
     bool m_useBgColor {false};
     bool m_useCustomBlendingOptions;
-    int m_customOpacity;
+    qreal m_customOpacity;
     QString m_customCompositeOp;
 };
